book-recommender-system/requirements.txt
scipy>=1.0.0
numpy>=1.15.1
pandas>=0.23.4
scikit_learn>=0.20.3
scikit_surprise>=1.0.6
scikit-optimize>=0.5.2
